% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimDementia.R
\name{estimDementia}
\alias{estimDementia}
\title{Computation of Health Indicators for dementia in France}
\usage{
estimDementia(t, nb_people = 100, nb_iter = 0, gender = "W")
}
\arguments{
\item{t}{year of the projections for health indicators.}

\item{nb_people}{number of people whose trajectory will be simulated for each generation. Default is \code{100}.}

\item{nb_iter}{number of iterations for the algorithm. Default is \code{0}.}

\item{gender}{gender for computation. \code{"W"} for women and \code{"M"} for men. Default is \code{"W"}.}
}
\value{
a list containing the health indicators
}
\description{
This function computes many health indicators without risk factor on transitions for a given year with all french data.
}
\examples{
estimDementia(t = 2040,
nb_people = 10000,
nb_iter = 100,
gender = "W")
}
