% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimHI.R
\name{estimHI}
\alias{estimHI}
\title{Computation of Health Indicators}
\usage{
estimHI(t, scenario, an_scenario, nbind, nb_iter, data_pop, sexe, data_a01,
  data_a02, data_theta01, data_theta02, data_theta12, data_conso, data_incid,
  data_rr_DvsND, data_a01_valeurs, data_a02_valeurs, data_theta01_valeurs,
  data_conso_valeurs, data_incid_valeurs, data_rr_DvsND_valeurs)
}
\arguments{
\item{t}{year for the projections.}

\item{scenario}{0 = pas de changement; 1 = reduction par 2 de la conso de benzo; 2 = reduction totale. Default is \code{0}.}

\item{an_scenario}{année de mise en place de la reduction de la consommation de benzo.}

\item{nbind}{nombre d'individus dont on va simuler la trajectoire pour chaque generation.}

\item{nb_iter}{nombre d'iterations de l'algo.}

\item{data_pop}{population en entree.}

\item{sexe}{sexe de la population en entree.}

\item{data_a01}{risques de devenir dement selon l'age.}

\item{data_a02}{risque de deces chez les non-dements par age et par annee.}

\item{data_theta01}{risque relatif de la demence selon la consommation de benzo.}

\item{data_theta02}{risque relatif de deces chez les non-dements selon la consommation de benzo.}

\item{data_theta12}{risque relatif de deces chez les dements selon la consommation de benzo.}

\item{data_conso}{prevalence de la consommation de benzo selon ag.}

\item{data_incid}{incidence de la consommation de benzo selon age.}

\item{data_rr_DvsND}{RR de deces pour les dements VS non-dements.}

\item{data_a01_valeurs}{variability of risques de devenir dement selon l'age.}

\item{data_a02_valeurs}{variability of risque de deces chez les non-dements par age et par annee.}

\item{data_theta01_valeurs}{variability of risque relatif de la demence selon la consommation de benzo.}

\item{data_conso_valeurs}{variability of prevalence de la consommation de benzo selon ag.}

\item{data_incid_valeurs}{variability of incidence de la consommation de benzo selon age.}

\item{data_rr_DvsND_valeurs}{variability of RR de deces pour les dements VS non-dements.}
}
\value{
a list containing the health indicators
}
\description{
This function computes many health indicators under several scenarios of intervention in benzodiazepines chronic consumption.
}
\examples{
estimHI(t = 2040,
scenario = 0,
an_scenario = 2020,
nbind = 10000,
nb_iter = 100,
data_pop = pop,
sexe = "F",
data_a01 = a01,
data_a02 = a02,
data_theta01 = theta01_cas_1_6,
data_theta02 = theta02_1,
data_theta12 = theta02_1,
data_conso = prevconso,
data_incid = incidconso,
data_rr_DvsND = rr_DvsND,
data_a01_valeurs = a01_valeurs,
data_a02_valeurs = a02_valeurs,
data_theta01_valeurs = theta01_cas_1_6_valeurs,
data_conso_valeurs <- prevconso_valeurs,
data_incid_valeurs <- incidconso_valeurs,
data_rr_DvsND_valeurs = rr_DvsND_valeurs)
}
