% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimHI_3FDR.R
\name{estimHI_3FDR}
\alias{estimHI_3FDR}
\title{Computation of Health Indicators}
\usage{
estimHI_3FDR(
  t,
  intervention = 0,
  year_intervention = NULL,
  nb_people = 100,
  data_pop,
  gender = "W",
  data_a01_values,
  data_a02_values,
  data_theta01_1_values,
  data_theta01_2_values,
  data_theta01_3_values,
  data_theta02_1_values,
  data_theta02_2_values,
  data_theta02_3_values,
  data_theta12_1_values,
  data_theta12_2_values,
  data_theta12_3_values,
  data_prev_0_values,
  data_prev_1_values,
  data_prev_2_values,
  data_prev_3_values,
  data_prev_4_values,
  data_prev_5_values,
  data_prev_6_values,
  data_prev_7_values,
  data_incid_0_values,
  data_incid_1_values,
  data_incid_3_values,
  data_incid_5_values,
  incid_global,
  theta1,
  theta3,
  data_rr_DvsND_values
)
}
\arguments{
\item{t}{year of the projections for health indicators.}

\item{intervention}{0 = no change; 1 = risk factor prevalence and incidence considered as null. Default is \code{0}.}

\item{year_intervention}{year of the intervention in risk factor distribution takes place. Default is \code{NULL}.}

\item{nb_people}{number of people whose trajectory will be simulated for each generation. Default is \code{100}.}

\item{data_pop}{data source for demographics data.}

\item{gender}{gender for computation. \code{"W"} for women and \code{"M"} for men. Default is \code{"W"}.}

\item{data_a01_values}{data source for the incidence of disease.}

\item{data_a02_values}{data source for the mortality of healthy subjects.}

\item{data_theta01_1_values}{data source for the relative risks associated with the exposure 1 for disease.}

\item{data_theta01_2_values}{data source for the relative risks associated with the exposure 2 for disease.}

\item{data_theta01_3_values}{data source for the relative risks associated with the exposure 3 for disease.}

\item{data_theta02_1_values}{data source for the relative risks associated with the exposure 1 for mortality among healthy subjects.}

\item{data_theta02_2_values}{data source for the relative risks associated with the exposure 2 for mortality among healthy subjects.}

\item{data_theta02_3_values}{data source for the relative risks associated with the exposure 3 for mortality among healthy subjects.}

\item{data_theta12_1_values}{data source for the relative risks associated with the exposure 1 for mortality among diseased subjects.}

\item{data_theta12_2_values}{data source for the relative risks associated with the exposure 2 for mortality among diseased subjects.}

\item{data_theta12_3_values}{data source for the relative risks associated with the exposure 3 for mortality among diseased subjects.}

\item{data_prev_0_values}{data source for the prevalence of the exposition 0.}

\item{data_prev_1_values}{data source for the prevalence of the exposition 1.}

\item{data_prev_2_values}{data source for the prevalence of the exposition 2.}

\item{data_prev_3_values}{data source for the prevalence of the exposition 3.}

\item{data_prev_4_values}{data source for the prevalence of the exposition 4.}

\item{data_prev_5_values}{data source for the prevalence of the exposition 5.}

\item{data_prev_6_values}{data source for the prevalence of the exposition 6.}

\item{data_prev_7_values}{data source for the prevalence of the exposition 7.}

\item{data_incid_0_values}{data source for the incidence of the exposition 0.}

\item{data_incid_1_values}{data source for the incidence of the exposition 1.}

\item{data_incid_3_values}{data source for the incidence of the exposition 3.}

\item{data_incid_5_values}{data source for the incidence of the exposition 5.}

\item{incid_global}{data source for the global diabete incidence.}

\item{theta1}{HR for HTA in the diabete risk}

\item{theta3}{HR for inact in the diabete risk}

\item{data_rr_DvsND_values}{data source for the relative risks associated with the disease for mortality.}
}
\value{
a list containing the health indicators
}
\description{
This function computes many health indicators under several scenarios of intervention in risk factor distribution for a given year.
}
