% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrenchDementia.R
\name{FrenchDementia}
\alias{FrenchDementia}
\title{Computation of Health Indicators for dementia in France}
\usage{
FrenchDementia(t, intervention = 0, year_intervention = NULL,
  nb_people = 100, nb_iter = 0, gender = "W")
}
\arguments{
\item{t}{year of the projections for health indicators.}

\item{intervention}{0 = no change; 1 = reduction by two of risk factor distribution; 2 = risk factor distribution considered as null. Default is \code{0}.}

\item{year_intervention}{year of the intervention in risk factor distribution takes place. Default is \code{NULL}.}

\item{nb_people}{number of people whose trajectory will be simulated for each generation. Default is \code{100}.}

\item{nb_iter}{number of iterations for the algorithm. Default is \code{0}.}

\item{gender}{gender for computation. \code{"W"} for women and \code{"M"} for men. Default is \code{"W"}.}
}
\value{
a list containing the health indicators
}
\description{
This function computes many health indicators under several interventions of intervention in risk factor distribution for a given year with all french data.
}
\examples{
FrenchDementia(t = 2040,
intervention = 1,
year_intervention = 2020,
nb_people = 10000,
nb_iter = 100,
gender = "W")
}
