% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiResultClass.R
\name{multiResultClass}
\alias{multiResultClass}
\title{Combination of multiple results}
\usage{
multiResultClass(LE_overall = NULL, LE_overall_exp = NULL,
  LE_overall_nonexp = NULL, LE_without_dis = NULL,
  LE_without_dis_exp = NULL, LE_without_dis_nonexp = NULL, LE_dis = NULL,
  LE_dis_exp = NULL, LE_dis_nonexp = NULL, LE_non_dis = NULL,
  LE_non_dis_exp = NULL, LE_non_dis_nonexp = NULL, np_age = NULL,
  tp_dis = NULL, nsurvival = NULL, rsurvival = NULL, nb_dis = NULL,
  nb_dis_exp = NULL, nb_dis_nonexp = NULL, p_dis = NULL, a_dis = NULL,
  m_exp = NULL, p_exp = NULL, mortality_r = NULL)
}
\arguments{
\item{LE_overall}{Default is \code{NULL}.}

\item{LE_overall_exp}{Default is \code{NULL}.}

\item{LE_overall_nonexp}{Default is \code{NULL}.}

\item{LE_without_dis}{Default is \code{NULL}.}

\item{LE_without_dis_exp}{Default is \code{NULL}.}

\item{LE_without_dis_nonexp}{Default is \code{NULL}.}

\item{LE_dis}{Default is \code{NULL}.}

\item{LE_dis_exp}{Default is \code{NULL}.}

\item{LE_dis_nonexp}{Default is \code{NULL}.}

\item{LE_non_dis}{Default is \code{NULL}.}

\item{LE_non_dis_exp}{Default is \code{NULL}.}

\item{LE_non_dis_nonexp}{Default is \code{NULL}.}

\item{np_age}{Default is \code{NULL}.}

\item{tp_dis}{Default is \code{NULL}.}

\item{nsurvival}{Default is \code{NULL}.}

\item{rsurvival}{Default is \code{NULL}.}

\item{nb_dis}{Default is \code{NULL}.}

\item{nb_dis_exp}{Default is \code{NULL}.}

\item{nb_dis_nonexp}{Default is \code{NULL}.}

\item{p_dis}{Default is \code{NULL}.}

\item{a_dis}{Default is \code{NULL}.}

\item{m_exp}{Default is \code{NULL}.}

\item{p_exp}{Default is \code{NULL}.}

\item{mortality_r}{Default is \code{NULL}.}
}
\value{
a list of the different results of the \code{varHI} function.
}
\description{
This function organizes the different results of the \code{varHI} function.
}
\examples{
multiResultClass()
}
