% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varHI.R
\name{varHI}
\alias{varHI}
\title{Computation of the variability of the Health Indicators}
\usage{
varHI(t, scenario, an_scenario, nbind, nb_iter, data_pop, sexe, an_proj,
  data_conso, data_incid, a010, a011, a01_global, a020, a021, a02_global, a120,
  a121, a12_global, data_a01, data_theta01, data_a02, data_theta02,
  data_theta12, RR, prb_dem, age_dem)
}
\arguments{
\item{t}{year for the projections.}

\item{scenario}{0 = pas de changement; 1 = reduction par 2 de la conso de benzo; 2 = reduction totale. Default is \code{0}.}

\item{an_scenario}{année de mise en place de la reduction de la consommation de benzo.}

\item{nbind}{nombre d'individus dont on va simuler la trajectoire pour chaque generation.}

\item{nb_iter}{nombre d'iterations de l'algo.}

\item{data_pop}{population en entree.}

\item{sexe}{sexe de la population en entree.}

\item{an_proj}{???}

\item{data_conso}{prevalence de la consommation de benzo selon ag.}

\item{data_incid}{incidence de la consommation de benzo selon age.}

\item{a010}{risques de devenir dement selon l'age.}

\item{a011}{risques de devenir dement selon l'age.}

\item{a01_global}{risques de devenir dement selon l'age.}

\item{a020}{risque de deces chez les non-dements par age et par annee.}

\item{a021}{risque de deces chez les non-dements par age et par annee.}

\item{a02_global}{risque de deces chez les non-dements par age et par annee.}

\item{a120}{risque de deces chez les dements par age et par annee.}

\item{a121}{risque de deces chez les dements par age et par annee.}

\item{a12_global}{risque de deces chez les dements par age et par annee.}

\item{data_a01}{risques de devenir dement selon l'age.}

\item{data_theta01}{risque relatif de la demence selon la consommation de benzo.}

\item{data_a02}{risque de deces chez les non-dements par age et par annee.}

\item{data_theta02}{risque relatif de deces chez les non-dements selon la consommation de benzo.}

\item{data_theta12}{risque relatif de deces chez les dements selon la consommation de benzo.}

\item{RR}{RR de deces pour les dements VS non-dements.}

\item{prb_dem}{??}

\item{age_dem}{??}
}
\value{
a list containing the variability of health indicators
}
\description{
This function computes many iterations of the \code{estimHI} function for compute the variability health indicators.
}
\examples{
varHI(t = t,
scenario = scenario,
an_scenario = an_scenario,
nbind = nbind,
nb_iter = nb_iter,
data_pop = data_pop,
sexe = sexe,
an_proj = an_proj,
data_conso = data_conso,
data_incid = data_incid,
a010 = a010,
a011 = a011,
a01_global = a01_global,
a020 = a020,
a021 = a021,
a02_global = a02_global,
a120 = a120,
a121 = a121,
a12_global = a12_global,
data_a01 = data_a01,
data_theta01 = data_theta01,
data_a02 = data_a02,
data_theta02 = data_theta02,
data_theta12 = data_theta12,
RR = RR,
prb_dem = prb_dem,
age_dem = age_dem)
}
