% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimHI.R
\name{estimHI}
\alias{estimHI}
\title{Computation of Health Indicators}
\usage{
estimHI(t, intervention = 0, year_intervention = NULL, nb_people = 100,
  nb_iter = 0, data_pop, gender = "W", data_a01_values, data_a02_values,
  data_theta01_values, data_theta02_values, data_theta12_values,
  data_prev_values, data_incid_values, data_rr_DvsND_values, data_a01, data_a02,
  data_theta01, data_theta02, data_theta12, data_prev, data_incid,
  data_rr_DvsND, Ncpus = 1)
}
\arguments{
\item{t}{year of the projections for health indicators.}

\item{intervention}{0 = no change; 1 = reduction by two of risk factor distribution; 2 = risk factor distribution considered as null. Default is \code{0}.}

\item{year_intervention}{year of the intervention in risk factor distribution takes place. Default is \code{NULL}.}

\item{nb_people}{number of people whose trajectory will be simulated for each generation. Default is \code{100}.}

\item{nb_iter}{number of iterations for the algorithm. Default is \code{0}.}

\item{data_pop}{data source for demographics data.}

\item{gender}{gender for computation. \code{"W"} for women and \code{"M"} for men. Default is \code{"W"}.}

\item{data_a01_values}{data source for the incidence of disease.}

\item{data_a02_values}{data source for the mortality of healthy subjects.}

\item{data_theta01_values}{data source for the relative risks associated with the exposure for disease.}

\item{data_theta02_values}{data source for the relative risks associated with the exposure for mortality among healthy subjects.}

\item{data_theta12_values}{data source for the relative risks associated with the exposure for mortality among diseased subjects.}

\item{data_prev_values}{data source for the prevalence of the exposition.}

\item{data_incid_values}{data source for the incidence of the exposition.}

\item{data_rr_DvsND_values}{data source for the relative risks associated with the disease for mortality.}

\item{data_a01}{variability of data source for the incidence of disease.}

\item{data_a02}{variability of data source for the mortality of healthy subjects.}

\item{data_theta01}{variability of data source for the relative risks associated with the exposure for disease.}

\item{data_theta02}{variability of data source for the relative risks associated with the exposure for mortality among healthy subjects.}

\item{data_theta12}{variability of data source for the relative risks associated with the exposure for mortality among diseased subjects.}

\item{data_prev}{variability of data source for the prevalence of the exposition.}

\item{data_incid}{variability of data source for the incidence of the exposition.}

\item{data_rr_DvsND}{variability of data source for the relative risks associated with the disease for mortality.}

\item{Ncpus}{The number of processors available. Default is \code{"1"}.}
}
\value{
a list containing the health indicators
}
\description{
This function computes many health indicators under several scenarios of intervention in risk factor distribution for a given year.
}
\examples{
estimHI(t = 2040,
intervention = 1,
year_intervention = 2020,
nb_people = 10000,
nb_iter = 100,
data_pop = pop,
gender = "W",
data_a01_values = a01_constant_values,
data_a02_values = a02_constant_values,
data_theta01_values = theta01_cas_1_6_values,
data_theta02_values = theta02_increase_values,
data_theta12_values = theta02_increase_values,
data_prev_values <- prevconso_values,
data_incid_values <- incidconso_values,
data_rr_DvsND_values = rr_DvsND_values,
data_a01 = a01_constant,
data_a02 = a02_constant,
data_theta01 = theta01_cas_1_6,
data_theta02 = theta02_increase,
data_theta12 = theta02_increase,
data_prev = prevconso,
data_incid = incidconso,
data_rr_DvsND = rr_DvsND,
Ncpus = 1)
}
