% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiResultClass.R
\name{multiResultClass}
\alias{multiResultClass}
\title{Combination of multiple results}
\usage{
multiResultClass(ev_gen = NULL, ev_gen_conso = NULL,
  ev_gen_nonconso = NULL, ev_sans_mal = NULL, ev_sans_mal_conso = NULL,
  ev_sans_mal_nonconso = NULL, ev_mal = NULL, ev_mal_conso = NULL,
  ev_mal_nonconso = NULL, ev_non_mal = NULL, ev_non_mal_conso = NULL,
  ev_non_mal_nonconso = NULL, np_age = NULL, tp_dem = NULL,
  nsurvie = NULL, tsurvie = NULL, nm_dem = NULL, nm_dem_conso = NULL,
  nm_dem_nonconso = NULL, p_dem = NULL, a_dem = NULL, m_conso = NULL,
  p_conso = NULL, q_mortalite = NULL)
}
\arguments{
\item{ev_gen}{Default is \code{NULL}.}

\item{ev_gen_conso}{Default is \code{NULL}.}

\item{ev_gen_nonconso}{Default is \code{NULL}.}

\item{ev_sans_mal}{Default is \code{NULL}.}

\item{ev_sans_mal_conso}{Default is \code{NULL}.}

\item{ev_sans_mal_nonconso}{Default is \code{NULL}.}

\item{ev_mal}{Default is \code{NULL}.}

\item{ev_mal_conso}{Default is \code{NULL}.}

\item{ev_mal_nonconso}{Default is \code{NULL}.}

\item{ev_non_mal}{Default is \code{NULL}.}

\item{ev_non_mal_conso}{Default is \code{NULL}.}

\item{ev_non_mal_nonconso}{Default is \code{NULL}.}

\item{np_age}{Default is \code{NULL}.}

\item{tp_dem}{Default is \code{NULL}.}

\item{nsurvie}{Default is \code{NULL}.}

\item{tsurvie}{Default is \code{NULL}.}

\item{nm_dem}{Default is \code{NULL}.}

\item{nm_dem_conso}{Default is \code{NULL}.}

\item{nm_dem_nonconso}{Default is \code{NULL}.}

\item{p_dem}{Default is \code{NULL}.}

\item{a_dem}{Default is \code{NULL}.}

\item{m_conso}{Default is \code{NULL}.}

\item{p_conso}{Default is \code{NULL}.}

\item{q_mortalite}{Default is \code{NULL}.}
}
\value{
a list of the different results of the \code{chi_mcmc} function.
}
\description{
This function organizes the different results of the \code{varHI} function.
}
\examples{
multiResultClass()
}
