% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varHI.R
\name{varHI}
\alias{varHI}
\title{Computation of the variability of the Health Indicators}
\usage{
varHI(t, intervention, year_intervention, nb_people, nb_iter, data_pop, gender,
  data_prev, data_incid, a010, a011, a01_global, a020, a021, a02_global, a120,
  a121, a12_global, data_a01, data_theta01, data_a02, data_theta02,
  data_theta12, RR, prb_dem, age_dem)
}
\arguments{
\item{t}{year of the projections for health indicators.}

\item{intervention}{0 = no change; 1 = reduction by two of risk factor distribution; 2 = risk factor distribution considered as null.}

\item{year_intervention}{year of the intervention in risk factor distribution takes place.}

\item{nb_people}{number of people whose trajectory will be simulated for each generation.}

\item{nb_iter}{number of iterations for the algorithm.}

\item{data_pop}{data source for demographics data.}

\item{gender}{gender for computation. "W" for women and "M" for men.}

\item{data_prev}{data source for the prevalence of the exposition.}

\item{data_incid}{data source for the incidence of the exposition.}

\item{a010}{incidence of disease on non exposed peoples.}

\item{a011}{incidence of disease on exposed peoples.}

\item{a01_global}{global incidence of disease.}

\item{a020}{mortality of healthy subjects on non exposed peoples.}

\item{a021}{mortality of healthy subjects on exposed peoples.}

\item{a02_global}{global mortality of healthy subjects.}

\item{a120}{mortality of diseased subjects on non exposed peoples.}

\item{a121}{mortality of diseased subjects on exposed peoples.}

\item{a12_global}{global mortality of diseased subjects.}

\item{data_theta01}{data source for the relative risks associated with the exposure for disease.}

\item{data_theta02}{data source for the relative risks associated with the exposure for mortality among healthy subjects.}

\item{data_theta12}{data source for the relative risks associated with the exposure for mortality among diseased subjects.}

\item{RR}{relative risks associated with the disease for mortality}

\item{prb_dem}{life-long probability of disease}

\item{age_dem}{average age at disease onset}
}
\value{
a list containing the variability of health indicators
}
\description{
This function computes many iterations of the \code{estimHI} function for compute the variability of the health indicators.
}
\examples{
varHI(t = t,
intervention = intervention,
year_intervention = year_intervention,
nb_people = nb_people,
nb_iter = nb_iter,
data_pop = data_pop,
gender = gender,
data_prev = data_prev,
data_incid = data_incid,
a010 = a010,
a011 = a011,
a01_global = a01_global,
a020 = a020,
a021 = a021,
a02_global = a02_global,
a120 = a120,
a121 = a121,
a12_global = a12_global,
data_theta01 = data_theta01,
data_theta02 = data_theta02,
data_theta12 = data_theta12,
RR = RR,
prb_dem = prb_dem,
age_dem = age_dem)
}
